+++ shop/tests/cart.py
@@ -5,6 +5,7 @@ from django.test.testcases import TestCase
 from shop.cart.modifiers_pool import cart_modifiers_pool
 from shop.models.cartmodel import Cart, CartItem
 from shop.models.productmodel import Product
+from shop.tests.util import Mock
 from shop.tests.utils.context_managers import SettingsOverride
 
 # This try except is there to let people run the tests from any project
@@ -22,8 +23,9 @@ class CartTestCase(TestCase):
 
     def setUp(self):
         cart_modifiers_pool.USE_CACHE = False
-        self.user = User.objects.create(username="test",
-            email="test@example.com")
+        user = User.objects.create(username="test", email="test@example.com")
+        self.request = Mock()
+        setattr(self.request, 'user', user)
         self.product = Product()
         self.product.name = "TestPrduct"
         self.product.slug = "TestPrduct"
@@ -36,13 +38,13 @@ class CartTestCase(TestCase):
         self.inactive_product.save()
 
         self.cart = Cart()
-        self.cart.user = self.user
+        self.cart.user = user
         self.cart.save()
 
     def test_empty_cart_costs_0_quantity_0(self):
         with SettingsOverride(SHOP_CART_MODIFIERS=[]):
 
-            self.cart.update()
+            self.cart.update(self.request)
 
             self.assertEqual(self.cart.subtotal_price, Decimal('0.0'))
             self.assertEqual(self.cart.total_price, Decimal('0.0'))
@@ -52,7 +54,7 @@ class CartTestCase(TestCase):
         with SettingsOverride(SHOP_CART_MODIFIERS=[]):
             self.cart.add_product(self.product)
             self.cart.save()
-            self.cart.update()
+            self.cart.update(self.request)
             self.cart.save()
 
             self.assertEqual(self.cart.subtotal_price, self.PRODUCT_PRICE)
@@ -64,7 +66,7 @@ class CartTestCase(TestCase):
 
             # We add two objects now :)
             self.cart.add_product(self.product, 2)
-            self.cart.update()
+            self.cart.update(self.request)
             self.cart.save()
 
             self.assertEqual(self.cart.subtotal_price, self.PRODUCT_PRICE * 2)
@@ -76,7 +78,7 @@ class CartTestCase(TestCase):
             'shop.cart.modifiers.tax_modifiers.TenPercentGlobalTaxModifier']
         with SettingsOverride(SHOP_CART_MODIFIERS=MODIFIERS):
             self.cart.add_product(self.product)
-            self.cart.update()
+            self.cart.update(self.request)
             self.cart.save()
 
             self.assertEqual(self.cart.subtotal_price, self.PRODUCT_PRICE)
@@ -90,7 +92,7 @@ class CartTestCase(TestCase):
         with SettingsOverride(SHOP_CART_MODIFIERS=MODIFIERS):
             self.cart.add_product(self.product)
 
-            self.cart.update()
+            self.cart.update(self.request)
             self.cart.save()
 
             self.assertEqual(self.cart.subtotal_price, self.PRODUCT_PRICE)
@@ -104,7 +106,7 @@ class CartTestCase(TestCase):
         with SettingsOverride(SHOP_CART_MODIFIERS=MODIFIERS):
             # We add 6 objects now :)
             self.cart.add_product(self.product, 6)
-            self.cart.update()
+            self.cart.update(self.request)
             self.cart.save()
 
             #subtotal is 600 - 10% = 540
@@ -122,7 +124,7 @@ class CartTestCase(TestCase):
             self.assertEqual(self.cart.total_quantity, 0)
             self.cart.add_product(self.product)
             self.cart.add_product(self.product)
-            self.cart.update()
+            self.cart.update(self.request)
             self.cart.save()
 
             self.assertEqual(len(self.cart.items.all()), 1)
@@ -134,7 +136,7 @@ class CartTestCase(TestCase):
             self.assertEqual(self.cart.total_quantity, 0)
             self.cart.add_product(self.product, merge=False)
             self.cart.add_product(self.product, merge=False)
-            self.cart.update()
+            self.cart.update(self.request)
             self.cart.save()
 
             self.assertEqual(len(self.cart.items.all()), 2)
@@ -156,7 +158,7 @@ class CartTestCase(TestCase):
             name="Variation 1", active=True)
         with SettingsOverride(SHOP_CART_MODIFIERS=[]):
             self.cart.add_product(variation)
-            self.cart.update()
+            self.cart.update(self.request)
             self.cart.save()
             self.assertEqual(self.cart.subtotal_price, self.PRODUCT_PRICE)
 
@@ -165,7 +167,7 @@ class CartTestCase(TestCase):
             self.assertEqual(self.cart.total_quantity, 0)
             self.cart.add_product(self.product)
             self.cart.add_product(self.product)
-            self.cart.update()
+            self.cart.update(self.request)
             self.cart.save()
 
             self.assertEqual(len(self.cart.items.all()), 1)
@@ -188,7 +190,7 @@ class CartTestCase(TestCase):
 
     def test_get_updated_cart_items(self):
         self.cart.add_product(self.product)
-        self.cart.update()
+        self.cart.update(self.request)
         cached_cart_items = self.cart.get_updated_cart_items()
 
         cart_items = CartItem.objects.filter(cart=self.cart)
@@ -206,6 +208,5 @@ class CartTestCase(TestCase):
     def test_adding_inactive_product_does_nothing(self):
         self.cart.add_product(self.product)
         self.cart.add_product(self.inactive_product)
-        self.cart.update()
+        self.cart.update(self.request)
         self.assertEqual(len(self.cart.items.all()), 1)
-