+++ shop/views/cart.py
@@ -112,11 +112,10 @@ class CartDetails(ShopTemplateResponseMixin, CartItemDetail):
     def get_context_data(self, **kwargs):
         # There is no get_context_data on super(), we inherit from the mixin!
         ctx = {}
-        state = {}
-        cart_object = get_or_create_cart(self.request)
-        cart_object.update(state)
-        ctx.update({'cart': cart_object})
-        ctx.update({'cart_items': cart_object.get_updated_cart_items()})
+        cart = get_or_create_cart(self.request)
+        cart.update(self.request)
+        ctx.update({'cart': cart})
+        ctx.update({'cart_items': cart.get_updated_cart_items()})
         return ctx
 
     def get(self, request, *args, **kwargs):