+++ shop/views/checkout.py
@@ -50,8 +50,8 @@ class CheckoutSelectionView(LoginMixin, ShopTemplateView):
         a reference to the Order on either the User object or the session.
         """
         cart = get_or_create_cart(self.request)
-        cart.update()
-        order = Order.objects.create_from_cart(cart)
+        cart.update(self.request)
+        order = Order.objects.create_from_cart(cart, self.request)
         request = self.request
         add_order_to_request(request, order)
         return order