+++ CHANGELOG.rst
@@ -4,7 +4,15 @@ Version NEXT
 * models.FloatField are now automatically localized.
 * Support for Django 1.2 and Django 1.3 dropped.
 * Product model now has property ``can_be_added_to_cart`` which is checked before adding the product to cart
-
+* In cart_modifiers methods ``get_extra_cart_price_field`` and ``get_extra_cart_item_price_field``
+  accepts the additional object ``request`` which can be used to calculate the price
+  according to the state of a session, the IP-address or whatever might be useful.
+  Note for backwards compatibility: Until version 0.1.2, instead of the ``request``
+  object, an empty Python dictionary named ``state`` was passed into the cart
+  modifiers. This ``state`` object could contain arbitrary data to exchange information
+  between the cart modifiers. This Python dict now is a temporary attribute of the
+  ``request`` object named ``cart_modifier_state``. Use it instead of the
+  ``state`` object.
 * Cart modifiers can add an optional ``data`` field beside ``label`` and ``value``
   for both, the ExtraOrderPriceField and the ExtraOrderItemPriceField model.
   This extra ``data`` field can contain anything serializable as JSON.